Require package immediately
Require in the middle allows us to hook into the package when the
package is actually required, so we don't have to load all the packages
at once.

When I was trying the library out in an express based app with `tsx`, I
noticed that the express hook wasn't fired (require-in-the-middle's
Hook), I think this is because `tsx` modifies the `require` function
to transpile TS files on the fly.

To deal with this, we should make sure express is required at all times.

Without request context we can't do anything.

I was first thinking about introducing an option to require the package
immediately but having options makes the code less readable and more
difficult to understand. Dropping an additional library is also a
good thing, then we're down to just one dependency.