Only apply hooks if there's an agent (and remove stat calls)
The sinks were responsible for calling onInspectedCall. Since we now
have a central place for the wrapping... We can simplify this by
reporting stats in the applyHooks function.

We should always have an agent before we apply hooks, there's no point
to intercept calls if there's no agent. This will also simplify testing
because code coverage will be easier.