Add attacks detected / blocked to stats (if running in dry mode)
When calling the interceptor, we can catch our own error and rethrow
that... then we know that we blocked an attack but what if running in
dry mode? We don't get to see the error... So we need to let the
interceptor return a result (which might be a detected attack). This
also simplifies the interceptor handling logic.