Reset lastIndex on regex object
Because we're using the `g` flag, the regex keeps track of `exec(...)`
calls. We need to reset it so that it starts from the beginning.

for (const match of matchAll(command, commandsRegex)) {
    // If the command doesn't contain the user input, we can skip this match
    if (userInput !== match[0]) {
      continue;
    }

Also, we need to use the actual match ^ here to check if that actual
command occurs is the userInput. Not just execute a regex again because
it might match another command from the regex.