Use environment variable to check if next is handling requests
In monorepo's with many packages, it could be that next is used in a
separate package (e.g. their website)

If we use `require.resolve("next")` it will return a path and thus start
reading bodies

For micro, there's no way to know if it's micro (`require("micro")` does not happen nor is there an env variable)

For micro, we can add a separate function handler to improve (similar to
lambda and google function).