Don't send heartbeat when there are compressed stats
We keep durations per operation (to measure the time our algo takes to
detect attacks)

After 5000 samples we compress them into percentiles. 5000 invocations
is not much. This means that we will send a heartbeat almost every
checkIfHeartbeatIsNeededEveryMS for normal production traffic

Which is currently set at 60s

So every minute we would send a heartbeat...