Let Prisma tests run using new instrumentation (including ESM)
To reset the loaded prisma client, we needed to `delete` entries in `require.cache`

This caused some weird state where the prisma client after deleting
wasn't being instrumented anymore...

We also couldn't run the tests on esm yet because we couldn't do the
delete trick.

So splitting the test files and using an exclusive file lock was the
best solution.